fix(serve): isolate runtime error formatting
