+++ packages/cli/src/serve/server/error-response.ts
@@ -17,6 +17,7 @@ import {
 } from '@qwen-code/qwen-code-core';
 import type { Response } from 'express';
 import { writeStderrLine } from '../../utils/stdioHelpers.js';
+import { errorMessage } from '../error-message.js';
 import {
   BranchWhilePromptActiveError,
   CancelSentinelCollisionError,
@@ -773,25 +774,7 @@ export function sendBridgeError(
   res.status(500).json(errorPayload(err));
 }
 
-/**
- * Coerce an arbitrary thrown value to a useful string. Plain `String(err)`
- * yields `[object Object]` for JSON-RPC-shaped errors (`{code, message,
- * data}`) which are exactly what the ACP SDK forwards from the agent. Try
- * the `message` field first, fall back to JSON-stringify, then `String`.
- */
-export function errorMessage(err: unknown): string {
-  if (err instanceof Error) return err.message;
-  if (err && typeof err === 'object') {
-    const maybe = (err as { message?: unknown }).message;
-    if (typeof maybe === 'string' && maybe.length > 0) return maybe;
-    try {
-      return JSON.stringify(err);
-    } catch {
-      /* fall through */
-    }
-  }
-  return String(err);
-}
+export { errorMessage } from '../error-message.js';
 
 /**
  * Build the JSON body for a 5xx response. The ACP SDK forwards
